fix pypath / use of pytest conventions around fixtures
