+++ .github/workflows/python-package-shared.yml
@@ -1,0 +1,78 @@
+name: Labelbox Python SDK Shared
+
+on:
+  workflow_call:
+    inputs:
+      python-version:
+        required: true
+        type: string
+      api-key:
+        required: true
+        type: string
+      da-test-key:
+        required: true
+        type: string
+      test-env:
+        required: true
+        type: string
+      fixture-profile: 
+        required: true
+        type: boolean
+jobs:
+  build:
+    runs-on: ubuntu-latest
+    env:
+      # make sure to tell tox to use these environs in tox.ini
+      LABELBOX_TEST_API_KEY: ${{ secrets[inputs.api-key] }}
+      DA_GCP_LABELBOX_API_KEY: ${{ secrets[inputs.da-test-key] }}
+      LABELBOX_TEST_ENVIRON: ${{ inputs.test-env }}
+      FIXTURE_PROFILE: ${{ inputs.fixture-profile }}
+    steps:
+    - name: Cancel previous workflow
+      uses: styfle/cancel-workflow-action@0.12.1
+      with:
+        access_token: ${{ github.token }}
+
+    - uses: actions/checkout@v4
+      with:
+        token: ${{ secrets.ACTIONS_ACCESS_TOKEN  }}
+        ref: ${{ github.head_ref }}
+
+    - name: Set up Python ${{ inputs.python-version }}
+      uses: actions/setup-python@v5
+      with:
+        python-version: ${{ inputs.python-version }}
+        cache: "pip"
+
+    - name: yapf
+      id: yapf
+      uses: AlexanderMelde/yapf-action@master
+      with:
+        args: --verbose --recursive --parallel --style "google"
+    - name: dependencies
+      run: |
+        sudo apt-get -y update
+        sudo apt install -y libsm6 \
+                          libxext6 \
+                            ffmpeg \
+                    libfontconfig1 \
+                      libxrender1 \
+                  libgl1-mesa-glx
+    - name: mypy
+      run: |
+        pip install strenum
+        pip install mypy==1.9.0
+        # TODO: consider installing requests typing
+        mypy -p labelbox --pretty --show-error-codes --non-interactive --install-types
+    - name: Install test dependencies
+      run: |
+        pip install tox==4.14.1
+    # TODO: replace tox.ini with what the Makefile does
+    # to make sure local testing is
+    # identical to github actions which uses tox.ini
+    - name: Test with tox (unit)
+      run: |
+        tox -e py -- tests/unit
+    - name: Test with tox (integration / data)
+      run: |
+        tox -e py -- tests/integration tests/data
