+++ pytest.ini
@@ -1,5 +1,5 @@
 [pytest]
 pythonpath = .
-addopts = -rP -vvv --reruns 5 --reruns-delay 10 --durations=20 -n 20
+addopts = -rP -vvv --reruns 5 --reruns-delay 10 --durations=20 -n 32
 markers =
     slow: marks tests as slow (deselect with '-m "not slow"')
