+++ tests/integration/test_ephemeral.py
@@ -1,13 +1,11 @@
 import os
 import pytest
-from support.integration_client import Environ, EphemeralClient, IntegrationClient
 
 
-@pytest.mark.skipif(
-    not os.environ.get('LABELBOX_TEST_ENVIRON') == Environ.EPHEMERAL.value,
-    reason='This test only runs in EPHEMERAL environment')
-def test_org_and_user_setup(client):
-    assert type(client) == EphemeralClient
+@pytest.mark.skipif(not os.environ.get('LABELBOX_TEST_ENVIRON') == 'ephemeral',
+                    reason='This test only runs in EPHEMERAL environment')
+def test_org_and_user_setup(client, ephmeral_client):
+    assert type(client) == ephmeral_client
     assert client.admin_client
     assert client.api_key != client.admin_client.api_key
 
@@ -17,8 +15,7 @@ def test_org_and_user_setup(client):
     assert user
 
 
-@pytest.mark.skipif(
-    os.environ.get('LABELBOX_TEST_ENVIRON') == Environ.EPHEMERAL.value,
-    reason='This test does not run in EPHEMERAL environment')
-def test_integration_client(client):
-    assert type(client) == IntegrationClient
+@pytest.mark.skipif(os.environ.get('LABELBOX_TEST_ENVIRON') == 'ephemeral',
+                    reason='This test does not run in EPHEMERAL environment')
+def test_integration_client(client, integration_client):
+    assert type(client) == integration_client
