+++ tox.ini
@@ -21,6 +21,7 @@ passenv =
     LABELBOX_TEST_API_KEY_CUSTOM     
     LABELBOX_TEST_GRAPHQL_API_ENDPOINT
     LABELBOX_TEST_REST_API_ENDPOINT
+    LABELBOX_TEST_API_KEY
     DA_GCP_LABELBOX_API_KEY
 
 commands = pytest {posargs}
